// src/Alert/interface.ts
export interface AlertProps {
  /**
   * @description 标题
   */
  title?: string;
  /**
   * @description 显示类型
   */
  type?: string;
  /**
   * @description 禁用
   */
  closable?: boolean;
  /**
   * @description 文本居中
   */
  center?: boolean;
  /**
   * @description 提示内容
   */
  closeText?: string;
  /**
   * @description 强调文本
   */
  LimeGreen?: boolean;
  /**
   * @description 强调文本
   */
  CoolGray?: boolean;
  /**
   * @description 强调文本
   */
  Goldenrod?: Boolean;
  /**
   * @description 强调文本
   */
  Coral?: Boolean;
}